refactor: update migrations para postgres

+++ backend/src/main/resources/db/migration/h2/V2__create_accounts.sql
@@ -3,7 +3,7 @@ CREATE TABLE accounts (
     agency INT NOT NULL,
     number INT NOT NULL,
     type VARCHAR(50) NOT NULL,
-    balance DOUBLE DEFAULT 0.0 NOT NULL,
+    balance numeric DEFAULT 0.0 NOT NULL,
     password VARCHAR NOT NULL,
     user_id UUID UNIQUE,
     status VARCHAR,
